+++ ifetch/cli.py
@@ -30,6 +30,13 @@ def main(argv=None):
             from .auth_cli import main as auth_main  # type: ignore
         return sys.exit(auth_main(argv[1:]))
 
+    if argv and argv[0] == 'snapshot':
+        if __package__ in (None, ""):
+            from ifetch.snapshot_cli import main as snap_main  # type: ignore
+        else:
+            from .snapshot_cli import main as snap_main  # type: ignore
+        return sys.exit(snap_main(argv[1:]))
+
     if argv and argv[0] == 'recover':
         if __package__ in (None, ""):
             from ifetch.recover_cli import main as recover_main  # type: ignore